+++ install/script/database.sql
@@ -5,8 +5,9 @@ DROP DATABASE IF EXISTS test;
 DELETE FROM mysql.db WHERE Db='test' OR Db='test\\_%';
 FLUSH PRIVILEGES;
 
-CREATE DATABASE kumoreg CHARACTER SET UTF8;
-CREATE DATABASE kumoreg_training CHARACTER SET UTF8;
+CREATE DATABASE IF NOT EXISTS kumoreg CHARACTER SET UTF8;
+CREATE DATABASE IF NOT EXISTS kumoreg_training CHARACTER SET UTF8;
+
 CREATE USER 'kumoreg'@'localhost' IDENTIFIED BY 'password';
 GRANT ALL PRIVILEGES ON kumoreg.* to 'kumoreg'@'localhost';
 CREATE USER 'kumoregtraining'@'localhost' IDENTIFIED BY 'password';